Add book issue fixed
